+++ src/main/java/com/canja/kutowerdefence/ui/EnemyView.java
@@ -3,5 +3,5 @@ package com.canja.kutowerdefence.ui;
 public class EnemyView {
     // imageview
     //has its own enemy
-    imageview
+    // imageview
 }