+++ changelog.md
@@ -1,3 +1,8 @@
 # Change Log
 
-Please list the changes you made here during developement
+Please list the changes you made here during development
+
+## 29.04
+### Atalay:
+- Created a basic gameplay view with map loading, very WIP
+- Some modifications to GameSession, MapSelectionOverlay, Map, and Routing for creating a gameplay screen with the selected map